+++ HISTORY.md
@@ -14,7 +14,7 @@ Based on RocksDB 8.1.1
 * AVX512: fix disabling other optimizations (#489)
 * stress test: fix decoding error (#498)
 * db_bench and stress: fix WBM initiation (#510)
-
+* Sanitize max_num_parallel_flushes in WBM if 0 (#460)
 ### Miscellaneous
 * disable failing unit tests and paired bloom filter stress testing
 * version: update Speedb patch version to 2.4.1 (#503)